Fix enemy alerts
TL;DR: raise enemies' alert emission and reception position above ground.
Long: Create a special alert origin - the origin of an outgoing alert and where an incoming alert will be received. It does not require its own object, any child object within the enemy with a fitting position can be used for it.