Implement teleportation between levels
Show Level Complete menu with the Continue button to continue to the next level.
It will continue to the forced next level (see `nextLevelOverride` field of gameManager) or the next level in sequence (if the levelId is not set to Other or equals LAST_LEVEL). If none of the mentioned is true, show